add grab mechanic
Player (the table) can now grab throw-able objects (throw not implemented yet) that are within are certain range. Rearranged the player asset to include an anchor parent object in hopes of stabilizing the position of a grabbed object, but that is still buggy.

The player is not equipped to deal with multiple grabs (e.g. cannot pick up an object, put it down, and pick it up again).